array of posts now passed from App to PostFeed
