+++ front-end/src/components/PostFeed/PostFeed.tsx
@@ -3,24 +3,11 @@ import "./PostFeed.css";
 import { Post } from '../../models/Post';
 import PostComponent from '../PostComponent/PostComonent';
 
-const posts: Post[] = [
-  {
-    id: 1,
-    username: "first_name",
-    avatar: "/testimage/avatar.jpeg",
-    image: "/testimage/mountain.jpeg",
-    description: "description 1!",
-  },
-  {
-    id: 2,
-    username: "second284",
-    avatar: "/testimage/man.jpeg",
-    image: "/testimage/bridge.jpeg",
-    description: "this sentence is a test",
-  },
-];
+interface PostFeedProps {
+  posts: Post[]
+}
 
-const Feed: React.FC = () => {
+const Feed: React.FC<PostFeedProps> = ({ posts }) => {
   return (
     <div className="feed">
       {posts.map((post) => (
